minor compile stuff. ifdef win32 the align in multithread.c

+++ relational.c
@@ -16,7 +16,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
- * $Id: relational.c,v 1.64 2009/03/25 10:07:40 c_anthon Exp $
+ * $Id: relational.c,v 1.65 2009/04/03 19:48:40 c_anthon Exp $
  */
 
 #include "config.h"
@@ -884,7 +884,7 @@ extern void CommandRelationalSelect(char *sz)
 			if (j > 0)
 				output(" | ");
 
-			outputf("%*s", rs->widths[j], rs->data[i][j]);
+			outputf("%*s", (int)rs->widths[j], rs->data[i][j]);
 		}
 		outputl("");
 	}
